+++ mex/backend/main.py
@@ -1,5 +1,6 @@
 from collections.abc import AsyncIterator
 from contextlib import asynccontextmanager
+from itertools import chain
 from typing import Any
 
 import uvicorn
@@ -8,6 +9,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.openapi.utils import get_openapi
 from pydantic import BaseModel
 
+from mex.backend.auxiliary.wikidata import router as wikidata_router
 from mex.backend.exceptions import handle_uncaught_exception
 from mex.backend.extracted.main import router as extracted_router
 from mex.backend.identity.main import router as identity_router
@@ -17,9 +19,10 @@ from mex.backend.merged.main import router as merged_router
 from mex.backend.rules.main import router as rules_router
 from mex.backend.security import has_read_access, has_write_access
 from mex.backend.settings import BackendSettings
+from mex.backend.transform import MExJSONResponse
 from mex.common.cli import entrypoint
 from mex.common.connector import CONNECTOR_STORE
-from mex.common.types import Identifier
+from mex.common.types import EXTRACTED_IDENTIFIER_CLASSES, MERGED_IDENTIFIER_CLASSES
 from mex.common.types.identifier import MEX_ID_PATTERN
 
 
@@ -45,12 +48,12 @@ def create_openapi_schema() -> dict[str, Any]:
         routes=app.routes,
         servers=[dict(url=settings.backend_api_url)],
     )
-    for subclass in Identifier.__subclasses__():
-        name = subclass.__name__
+    for identifier in chain(EXTRACTED_IDENTIFIER_CLASSES, MERGED_IDENTIFIER_CLASSES):
+        name = identifier.__name__
         openapi_schema["components"]["schemas"][name] = {
             "title": name,
             "type": "string",
-            "description": subclass.__doc__,
+            "description": identifier.__doc__,
             "pattern": MEX_ID_PATTERN,
         }
 
@@ -77,6 +80,7 @@ app = FastAPI(
         email="mex@rki.de",
         url="https://github.com/robert-koch-institut/mex-backend",
     ),
+    default_response_class=MExJSONResponse,
     lifespan=lifespan,
     version="v0",
 )
@@ -87,6 +91,7 @@ router.include_router(identity_router, dependencies=[Depends(has_write_access)])
 router.include_router(ingest_router, dependencies=[Depends(has_write_access)])
 router.include_router(merged_router, dependencies=[Depends(has_read_access)])
 router.include_router(rules_router, dependencies=[Depends(has_write_access)])
+router.include_router(wikidata_router, dependencies=[Depends(has_read_access)])
 
 
 class SystemStatus(BaseModel):