+++ CHANGELOG.md
@@ -19,6 +19,40 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.18.0] - 2024-08-05
+
+### Changes
+
+- BulkIngestRequest contains now one single list "items"
+- tests for ingestion adapted to BulkIngestRequest-Model
+- remove stop-gaps for MX-1596
+
+### Removed
+- removed class  _BaseBulkIngestRequest for ingestion model
+
+## [0.17.0] - 2024-07-29
+
+### Changes
+
+- BREAKING: use `items` instead of `results` in paged wiki response
+- downgrade query logging to log level `debug`
+
+### Fixed
+
+- fix endpoint serializing not working with `mex.common.types`
+
+## [0.16.0] - 2024-07-26
+
+### Added
+
+- `/wikidata` endpoint to fetch all matching organizations from wikidata
+- add support for computed fields in graph queries
+
+### Changes
+
+- BREAKING: make `MEX_EXTRACTED_PRIMARY_SOURCE` an instance of its own class
+  instead of ExtractedPrimarySource in order to set static provenance identifiers
+
 ## [0.15.0] - 2024-06-14
 
 ### Added